+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.O365OrgSettings.Tests.ps1
@@ -45,6 +45,23 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     displayConcealedNames = $true
                 }
             }
+
+            Mock -CommandName Get-M365DSCOrgSettingsInstallationOptions -MockWith {
+                return @{
+                    '@odata.context' = 'https://graph.microsoft.com/beta/$metadata#admin/microsoft365Apps/installationOptions/$entity'
+                    updateChannel = 'current'
+                    appsForMac = @{
+                        isSkypeForBusinessEnabled = $True
+                        isMicrosoft365AppsEnabled = $true
+                    }
+                    appsForWindows = @{
+                        isVisioEnabled            = $True
+                        isSkypeForBusinessEnabled = $False
+                        isMicrosoft365AppsEnabled = $true
+                        isProjectEnabled          = $true
+                    }
+                }
+            }
         }
 
         # Test contexts
@@ -54,6 +71,9 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     AdminCenterReportDisplayConcealedNames     = $True;
                     IsSingleInstance                           = 'Yes'
                     M365WebEnableUsersToOpenFilesFrom3PStorage = $False;
+                    InstallationOptionsAppsForMac              = @('isSkypeForBusinessEnabled', 'isMicrosoft365AppsEnabled')
+                    InstallationOptionsAppsForWindows          = @('isVisioEnabled', 'isMicrosoft365AppsEnabled', 'isProjectEnabled')
+                    InstallationOptionsUpdateChannel           = 'current'
                     Ensure                                     = 'Present'
                     Credential                                 = $Credential
                 }